:truck: Chore(find): hit->find.c
rename hit.c -> find.c
